Walls, Basic Enemy, Projectile
Added walls that summon blocks to simmulate movement. Not 100% sure I like them, but left them in scene.

Added a basic enemy that can be hit by the basic projectiles with health.

Changed the size and location of player so they can only move horizontally. Additionally, limited the fire speed of the player so they cannot spam the projectile.