PowerUp, Constant Fire, Enemy
Added a powerup template and a 2x fire rate powerup. I plan to have this appear as a choose one of two scrolling down from the top of the screen.

The player now constantly fires without having to click a button. This improves ease of playability.

Enemy does not spin when hit with a projectile